Use vertex pulling

+++ src/block/BlockModel.java
@@ -2,8 +2,8 @@ package block;
 
 public class BlockModel {
     // depth, left, top, width, height, texture id
-    public short[][][] faces;
-    public BlockModel(short[][][] faces) {
+    public byte[][][] faces;
+    public BlockModel(byte[][][] faces) {
         this.faces = faces;
     }
 }
