+++ src/block/SlabBlock.java
@@ -12,7 +12,7 @@ public class SlabBlock extends Block {
 
     @Override
     public BlockModel getModel(BlockState blockState) {
-        return new BlockModel(new short[][][] {
+        return new BlockModel(new byte[][][] {
                 {{0, 0, 0, 16, 8, 0}},
                 {{0, 0, 0, 16, 8, 0}},
                 {{0, 0, 0, 16, 8, 0}},
